ai-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
Pillow==10.0.1
numpy==1.24.3
python-multipart==0.0.6
